backend/requirements.txt
boto3>=1.34.0
botocore>=1.34.0
flask>=2.3.0
flask-cors>=4.0.0
Pillow>=10.0.0
